Converting Long latency to Int latency - first commit (errors in runtime - need to adjust tap rdd buffers to also store int)

+++ core/src/main/scala/org/apache/spark/lineage/perfdebug/ignite/perftrace/IgniteCacheAggregateStatsStorage.scala
@@ -41,7 +41,7 @@ class IgniteCacheAggregateStatsStorage(ignite: Ignite = {
   
   private def deserializeStats(value: CacheValue): AggregateLatencyStats = {
     val split = value.split(separator, 3)
-    AggregateLatencyStats(split(0).toLong, split(1).toLong, split(2).toLong)
+    AggregateLatencyStats(split(0).toLong, split(1).toLong, split(2).toInt) // last one is latency
   }
   
   private def buildAggStatsCacheName(appId: String) = {
